Fix UTF-8 string marshaling between libobs and wrapper.
Normally, strings are marshaled to ANSI-strings by default, which
results in lost or badly formatted characters. This caused some issues
in cases where string comparison happens (for example: window capture)
after data has been passed through the wrapper and back.

This also gets rid of the ugly rewrapped method calls, replacing them
with even more ugly and long attributes.

+++ libobs-sharp/src/ObsData.cs
@@ -184,7 +184,7 @@ namespace OBS
 
 		public unsafe void SetInt(string name, int val)
 		{
-			libobs.obs_data_set_int((IntPtr)instance, name, val);
+			libobs.obs_data_set_int((IntPtr)instance, name, (long)val);
 		}
 
 		public unsafe void SetDouble(string name, double val)
